// condoplanner-app/src/apiClient/models/CommonAreaDto.ts
export interface CommonAreaDto {
    /**
     * 
     * @type {number}
     * @memberof CommonAreaDto
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof CommonAreaDto
     */
    condominiumId?: number;
    /**
     * 
     * @type {string}
     * @memberof CommonAreaDto
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CommonAreaDto
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CommonAreaDto
     */
    type?: string | null;
    /**
     * 
     * @type {number}
     * @memberof CommonAreaDto
     */
    capacity?: number;
    /**
     * 
     * @type {string}
     * @memberof CommonAreaDto
     */
    openingTime?: string;
    /**
     * 
     * @type {string}
     * @memberof CommonAreaDto
     */
    closingTime?: string;
    /**
     * 
     * @type {number}
     * @memberof CommonAreaDto
     */
    maxDuration?: number;
    /**
     * 
     * @type {boolean}
     * @memberof CommonAreaDto
     */
    available?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof CommonAreaDto
     */
    requiresApproval?: boolean;
    /**
     * 
     * @type {number}
     * @memberof CommonAreaDto
     */
    availableDays?: number;
    /**
     * 
     * @type {string}
     * @memberof CommonAreaDto
     */
    notes?: string | null;
    /**
     * 
     * @type {Array<CommonAreaPhotoDto>}
     * @memberof CommonAreaDto
     */
    photos?: Array<CommonAreaPhotoDto> | null;
}