// condoplanner-app/src/apiClient/models/UploadCommonAreaPhotoInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UploadCommonAreaPhotoInput
 */
export interface UploadCommonAreaPhotoInput {
    /**
     * 
     * @type {number}
     * @memberof UploadCommonAreaPhotoInput
     */
    areaId?: number;
    /**
     * 
     * @type {string}
     * @memberof UploadCommonAreaPhotoInput
     */
    contentType?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UploadCommonAreaPhotoInput
     */
    fileName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UploadCommonAreaPhotoInput
     */
    base64Original?: string | null;
}

/**
 * Check if a given object implements the UploadCommonAreaPhotoInput interface.
 */
export function instanceOfUploadCommonAreaPhotoInput(value: object): value is UploadCommonAreaPhotoInput {
    return true;
}

export function UploadCommonAreaPhotoInputFromJSON(json: any): UploadCommonAreaPhotoInput {
    return UploadCommonAreaPhotoInputFromJSONTyped(json, false);
}

export function UploadCommonAreaPhotoInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): UploadCommonAreaPhotoInput {
    if (json == null) {
        return json;
    }
    return {
        
        'areaId': json['areaId'] == null ? undefined : json['areaId'],
        'contentType': json['contentType'] == null ? undefined : json['contentType'],
        'fileName': json['fileName'] == null ? undefined : json['fileName'],
        'base64Original': json['base64Original'] == null ? undefined : json['base64Original'],
    };
}

export function UploadCommonAreaPhotoInputToJSON(json: any): UploadCommonAreaPhotoInput {
    return UploadCommonAreaPhotoInputToJSONTyped(json, false);
}

export function UploadCommonAreaPhotoInputToJSONTyped(value?: UploadCommonAreaPhotoInput | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'areaId': value['areaId'],
        'contentType': value['contentType'],
        'fileName': value['fileName'],
        'base64Original': value['base64Original'],
    };
}